+++ starkbank/split/__init__.py
@@ -1,0 +1,3 @@
+from .__split import get, query, page
+from .log.__log import Log
+from . import log